requirements.txt
streamlit
fpdf2
qrcode
pillow
pandas
numpy